+++ src/sql/__init__.py
@@ -1,5 +1,5 @@
-from .magic import RenderMagic, SqlMagic, load_ipython_extension
-from .connection import PLOOMBER_DOCS_LINK_STR
+from sql.magic import RenderMagic, SqlMagic, load_ipython_extension
+from sql.connection import PLOOMBER_DOCS_LINK_STR
 
 __version__ = "0.9.0dev"
